Split personal namespace into Safe and Unsafe part
