+++ rpc/src/v1/impls/mod.rs
@@ -32,6 +32,7 @@ mod ethcore;
 mod ethcore_set;
 mod net;
 mod personal;
+mod personal_accounts;
 mod personal_signer;
 mod rpc;
 mod traces;
@@ -43,6 +44,7 @@ pub use self::eth_filter::EthFilterClient;
 pub use self::eth_signing::{EthSigningUnsafeClient, EthSigningQueueClient};
 pub use self::net::NetClient;
 pub use self::personal::PersonalClient;
+pub use self::personal_accounts::PersonalAccountsClient;
 pub use self::personal_signer::SignerClient;
 pub use self::ethcore::EthcoreClient;
 pub use self::ethcore_set::EthcoreSetClient;
