+++ rpc/src/v1/tests/mocked/personal.rs
@@ -19,7 +19,7 @@ use std::str::FromStr;
 use jsonrpc_core::IoHandler;
 use util::{U256, Uint, Address};
 use ethcore::account_provider::AccountProvider;
-use v1::{PersonalClient, Personal};
+use v1::{PersonalClient, PersonalAccountsClient, PersonalAccounts, Personal};
 use v1::tests::helpers::TestMinerService;
 use ethcore::client::TestBlockChainClient;
 use ethcore::transaction::{Action, Transaction};
@@ -50,10 +50,12 @@ fn setup() -> PersonalTester {
 	let accounts = accounts_provider();
 	let client = blockchain_client();
 	let miner = miner_service();
-	let personal = PersonalClient::new(&accounts, &client, &miner, false);
+	let personal = PersonalClient::new(&accounts, &client);
+	let personal_accounts = PersonalAccountsClient::new(&accounts, &client, &miner, false);
 
 	let io = IoHandler::new();
 	io.add_delegate(personal.to_delegate());
+	io.add_delegate(personal_accounts.to_delegate());
 
 	let tester = PersonalTester {
 		accounts: accounts,
