+++ rpc/src/v1/traits/mod.rs
@@ -30,7 +30,7 @@ pub use self::web3::Web3;
 pub use self::eth::{Eth, EthFilter};
 pub use self::eth_signing::EthSigning;
 pub use self::net::Net;
-pub use self::personal::{Personal, PersonalSigner};
+pub use self::personal::{Personal, PersonalAccounts, PersonalSigner};
 pub use self::ethcore::Ethcore;
 pub use self::ethcore_set::EthcoreSet;
 pub use self::traces::Traces;
