+++ rpc/src/v1/traits/personal.rs
@@ -18,11 +18,31 @@
 use std::sync::Arc;
 use jsonrpc_core::*;
 
-/// Personal rpc interface.
+/// Personal rpc interface. Safe (read-only) functions.
 pub trait Personal: Sized + Send + Sync + 'static {
 
 	/// Lists all stored accounts
 	fn accounts(&self, _: Params) -> Result<Value, Error>;
+
+	/// Returns accounts information.
+	fn accounts_info(&self, _: Params) -> Result<Value, Error>;
+
+	/// Returns the accounts available for importing from Geth.
+	fn geth_accounts(&self, _: Params) -> Result<Value, Error>;
+
+	/// Should be used to convert object to io delegate.
+	fn to_delegate(self) -> IoDelegate<Self> {
+		let mut delegate = IoDelegate::new(Arc::new(self));
+		delegate.add_method("personal_listAccounts", Personal::accounts);
+		delegate.add_method("personal_accountsInfo", Personal::accounts_info);
+		delegate.add_method("personal_listGethAccounts", Personal::geth_accounts);
+
+		delegate
+	}
+}
+
+/// Personal rpc methods altering stored accounts or their settings.
+pub trait PersonalAccounts: Sized + Send + Sync + 'static {
 
 	/// Creates new account (it becomes new current unlocked account)
 	/// Param is the password for the account.
@@ -56,31 +76,22 @@ pub trait Personal: Sized + Send + Sync + 'static {
 	/// Set an account's metadata string.
 	fn set_account_meta(&self, _: Params) -> Result<Value, Error>;
 
-	/// Returns accounts information.
-	fn accounts_info(&self, _: Params) -> Result<Value, Error>;
-
-	/// Returns the accounts available for importing from Geth.
-	fn geth_accounts(&self, _: Params) -> Result<Value, Error>;
-
 	/// Imports a number of Geth accounts, with the list provided as the argument.
 	fn import_geth_accounts(&self, _: Params) -> Result<Value, Error>;
 
 	/// Should be used to convert object to io delegate.
 	fn to_delegate(self) -> IoDelegate<Self> {
 		let mut delegate = IoDelegate::new(Arc::new(self));
-		delegate.add_method("personal_listAccounts", Personal::accounts);
-		delegate.add_method("personal_newAccount", Personal::new_account);
-		delegate.add_method("personal_newAccountFromPhrase", Personal::new_account_from_phrase);
-		delegate.add_method("personal_newAccountFromWallet", Personal::new_account_from_wallet);
-		delegate.add_method("personal_unlockAccount", Personal::unlock_account);
-		delegate.add_method("personal_testPassword", Personal::test_password);
-		delegate.add_method("personal_changePassword", Personal::change_password);
-		delegate.add_method("personal_signAndSendTransaction", Personal::sign_and_send_transaction);
-		delegate.add_method("personal_setAccountName", Personal::set_account_name);
-		delegate.add_method("personal_setAccountMeta", Personal::set_account_meta);
-		delegate.add_method("personal_accountsInfo", Personal::accounts_info);
-		delegate.add_method("personal_listGethAccounts", Personal::geth_accounts);
-		delegate.add_method("personal_importGethAccounts", Personal::import_geth_accounts);
+		delegate.add_method("personal_newAccount", PersonalAccounts::new_account);
+		delegate.add_method("personal_newAccountFromPhrase", PersonalAccounts::new_account_from_phrase);
+		delegate.add_method("personal_newAccountFromWallet", PersonalAccounts::new_account_from_wallet);
+		delegate.add_method("personal_unlockAccount", PersonalAccounts::unlock_account);
+		delegate.add_method("personal_testPassword", PersonalAccounts::test_password);
+		delegate.add_method("personal_changePassword", PersonalAccounts::change_password);
+		delegate.add_method("personal_signAndSendTransaction", PersonalAccounts::sign_and_send_transaction);
+		delegate.add_method("personal_setAccountName", PersonalAccounts::set_account_name);
+		delegate.add_method("personal_setAccountMeta", PersonalAccounts::set_account_meta);
+		delegate.add_method("personal_importGethAccounts", PersonalAccounts::import_geth_accounts);
 
 		delegate
 	}
@@ -108,6 +119,7 @@ pub trait PersonalSigner: Sized + Send + Sync + 'static {
 		delegate.add_method("personal_confirmRequest", PersonalSigner::confirm_request);
 		delegate.add_method("personal_rejectRequest", PersonalSigner::reject_request);
 		delegate.add_method("personal_generateAuthorizationToken", PersonalSigner::generate_token);
+
 		delegate
 	}
 }
